refactor: addd useBookDetails hook
